+++ .github/workflows/editorconfig.yml
@@ -44,7 +44,7 @@ jobs:
 
       - name: resharper-cleanup
         run: |                    
-          jb cleanupcode Consolonia.sln
+          jb cleanupcode Consolonia.sln --exclude="**Consolonia.GuiCS\**.*"
         working-directory: ./src
 
       - name: Commit files
