+++ src/Consolonia.Core/ApplicationStartup.cs
@@ -24,14 +24,6 @@ namespace Consolonia.Core
             lifetime.Start(args);
         }
 
-        // ReSharper disable once UnusedParameter.Global
-        // ReSharper disable once UnusedMember.Global
-        public static TAppBuilder UseAutoDetectTerminal<TAppBuilder>(this TAppBuilder builder)
-            where TAppBuilder : AppBuilderBase<TAppBuilder>, new()
-        {
-            throw new NotImplementedException();
-        }
-
         public static TAppBuilder UseStandardConsole<TAppBuilder>(this TAppBuilder builder)
             where TAppBuilder : AppBuilderBase<TAppBuilder>, new()
         {
