+++ src/Consolonia.Gallery/Consolonia.Gallery.csproj
@@ -7,13 +7,14 @@
 
     <ItemGroup>
       <ProjectReference Include="..\Consolonia.Core\Consolonia.Core.csproj" />
+      <ProjectReference Include="..\Consolonia.PlatformSupport\Consolonia.PlatformSupport.csproj" />
       <ProjectReference Include="..\Consolonia.Themes.TurboVision\Consolonia.Themes.TurboVision.csproj" />
     </ItemGroup>
 
     <ItemGroup>
       <PackageReference Include="Avalonia" Version="$(AvaloniaVersion)" />
       <PackageReference Include="Avalonia.Diagnostics" Version="$(AvaloniaVersion)" />
-      <PackageReference Include="Avalonia.ReactiveUI" Version="$(AvaloniaVersion)"/>
+      <PackageReference Include="Avalonia.ReactiveUI" Version="$(AvaloniaVersion)" />
     </ItemGroup>
 
 </Project>
