+++ src/Consolonia.Gallery/Program.cs
@@ -1,5 +1,7 @@
 using System;
+using Avalonia;
 using Consolonia.Core;
+using Consolonia.PlatformSupport;
 
 namespace Consolonia.Gallery
 {
@@ -9,8 +11,13 @@ namespace Consolonia.Gallery
         private static void Main(string[] args)
         {
             if (args.Length > 1) throw new NotSupportedException();
-
-            ApplicationStartup.StartConsolonia<App>(args);
+            
+            AppBuilder.Configure<App>()
+                .UseConsolonia()
+                .UseAutoDetectedConsole()
+                //.UseConsole(new CursesConsole())
+                .LogToTrace()
+                .StartWithConsoleLifetime(args);
         }
     }
 }
