+++ src/Consolonia.GuiCS/.editorconfig
