+++ src/Consolonia.GuiCS/Assembly.cs
@@ -1,0 +1,7 @@
+using System;
+using System.Runtime.CompilerServices;
+
+[assembly: InternalsVisibleTo("Consolonia.TestsCore")]
+[assembly: InternalsVisibleTo("Consolonia.PlatformSupport")]
+
+[assembly: CLSCompliant(false)] //todo: should we make it compliant?
