+++ src/Consolonia.GuiCS/Consolonia.GuiCS.csproj
@@ -1,0 +1,10 @@
+<Project Sdk="Microsoft.NET.Sdk">
+    <ItemGroup>
+        <PackageReference Include="NStack.Core" Version="0.16.0" />
+    </ItemGroup>
+    <PropertyGroup>        
+        <WarningsAsErrors>false</WarningsAsErrors>
+        <EnforceCodeStyleInBuild>false</EnforceCodeStyleInBuild>
+        <EnableNETAnalyzers>false</EnableNETAnalyzers>
+    </PropertyGroup>
+</Project>
