+++ src/Consolonia.GuiCS/UnmanagedLibrary.cs
