+++ src/Consolonia.GuiCS/binding.cs
