+++ src/Consolonia.GuiCS/constants.cs
