+++ src/Consolonia.GuiCS/readme.md
@@ -1,0 +1,7 @@
+Almost all these files are either original or partly copypasted from https://github.com/migueldeicaza/gui.cs
+
+We need to think how to re-use them (ask to extract and pack within gui.cs or just make it mergable).
+We also need to check lincensing question as probably current library is more open.
+
+
+Autoformatting is switch at the CI for this project
