+++ src/Consolonia.PlatformSupport/Assembly.cs
